+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.5.2
 
 - Fix newer Forge installs not working on linux due to bad logic (#386)
+- Fix OSX cut/copy/paste shortcuts not working for all fields (#387)